Add webhint as local running script
Since webhint retired their online scanner, we need to add it as a local
devDependency and run it scanner as a node script. We'll use the html
formatter and the jsdom connector and then, after the HTML is generated,
we use pupeteer to generate the PDF from that, cleaning up after
ourselves.

At the moment, webhint has not PDF formatter.

related to: #76