Implement simpler check for analyzing the correct URL
Instead of storing an additional global, such as originalUrl, we can
rely on a try/catch block to check whether the domain supports https or
not and then let webhint analyze it. We first try https and if that
fails, we fall back to http. If that fails as well, then it must not be
a valid URL.

related to #68